+++ packages/@tinacms/alerts/README.md
@@ -1,0 +1,1 @@
+# @tinacms/alerts
