+++ packages/@tinacms/alerts/package.json
@@ -1,0 +1,35 @@
+{
+  "name": "@tinacms/alerts",
+  "version": "0.0.0",
+  "main": "build/index.js",
+  "types": "build/index.d.ts",
+  "license": "Apache-2.0",
+  "files": [
+    "build"
+  ],
+  "keywords": [
+    "tinacms",
+    "cms"
+  ],
+  "bugs": {
+    "url": "https://github.com/tinacms/tinacms/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/tinacms/tinacms.git",
+    "directory": "packages/@tinacms/alerts"
+  },
+  "scripts": {
+    "test": "jest --passWithNoTests",
+    "watch": "tinacms-scripts watch",
+    "dev": "tinacms-scripts dev",
+    "docs": "npx typedoc",
+    "build": "tinacms-scripts build"
+  },
+  "dependencies": {
+    "@tinacms/core": "0.7.2"
+  },
+  "devDependencies": {
+    "@tinacms/scripts": "^0.1.13"
+  }
+}
