+++ packages/tinacms/package.json
@@ -35,10 +35,11 @@
     "styled-components": ">=4.1"
   },
   "dependencies": {
-    "@tinacms/core": "^0.7.3",
-    "@tinacms/fields": "^0.8.0",
-    "@tinacms/form-builder": "^0.3.0",
-    "@tinacms/forms": "^0.3.0",
+    "@tinacms/alerts": "^0.0.0",
+    "@tinacms/core": "^0.7.2",
+    "@tinacms/fields": "^0.7.0",
+    "@tinacms/form-builder": "^0.2.16",
+    "@tinacms/forms": "^0.2.1",
     "@tinacms/icons": "^0.7.0",
     "@tinacms/media": "^0.0.0",
     "@tinacms/react-core": "^0.2.9",
